Split to small section
